serverBridge/requirements.txt
# Dipendenze necessarie per il bridge MQTT/WebSocket e l'API Flask

# Web Server Framework
Flask==3.0.3 

# WebSocket Server (per la Dashboard)
websockets==12.0

# MQTT Client (per le schede ESP32)
paho-mqtt==1.6.1

# Database Connector (per l'API Python/MySQL)
mysql-connector-python==8.4.0

# Librerie di supporto
Werkzeug==3.0.3  # Necessario per Flask
itsdangerous==2.1.2 # Necessario per Flask
Jinja2==3.1.4 # Necessario per Flask
MarkupSafe==2.1.5 # Necessario per Flask